pyproject.toml
[project]
name = "ryze-data"
version = "1.0.0"
description = "Comprehensive data processing framework for scientific papers"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "Pillow>=10.0.0",
    "pypdf>=3.0.0",
    "pdf2image>=1.16.0",
    "openai>=1.0.0",
    "backoff>=2.2.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "pydantic>=2.0.0",
    "jsonschema>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

deepseek-ocr = [
    "torch>=2.0.0",
    "transformers==4.46.3",
    "flash-attn>=2.0.0",
    "einops>=0.7.0",
    "addict>=2.4.0",
    "easydict>=1.10",
]

markitdown = ["markitdown>=0.1.0"]

benchmark = [
    "datasets>=2.14.0",
]

[project.scripts]
ryze-data = "src.cli.main:main"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
]
